requirementMLR.txt
numpy
pandas
streamlit
sklearn.linear_model=LinearRegression
base64